refactor: extend ChannelWallet instead of createing a new browser wallet
